+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Improve First Region creation. Check if main camera is perspective or orthographic
+- Improve Create Regions GUI. Inspector and Overlay Scene show the same extra GUI helpers
+- Fix Create New Regions path button. It opens on the local project Scene folder
 
 ## [1.0.0-preview.0] - 2020-10-24
 ### Add